pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "pandas>=2.3.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "redis>=6.4.0",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
]
